Split tests for Response and RedirectResponse
